plugins/python: Fix handling body passed to uwsgi.spool()
When one uses uwsgi.spool(body=<body>), the keyword arguments dict
(kw) will have "body" (string) as a key, not b"body" (bytes).

So, using PyBytes_FromString to get the key is wrong. We can use
the PyDict_GetItemString / PyDict_DelItemString helper functions
which do the right thing: construct a Python string from a char *
pointer, and use that string as a lookup key. This removes the
difference between Python 2 and Python 3 code, so the custom macros
are not needed anymore.

This also fixes a memory leak: PyBytes_FromString returned a new
reference which needed to be DECREF'ed, but this did not happen.